ROD-92 combine both custom validations into 1 file
