+++ apps/rod/translations/src/en/pages.json
@@ -59,9 +59,6 @@
       "date-of-application": {
         "label": "Date of submission"
       },
-      "cancel-application": {
-        "label": "Cancelling application?"
-      },
       "dependant-or-guardian": {
         "label": "Sponsor type"
       },
